requirements.txt
paramiko
python-frontmatter
pyyaml
tabulate
pypandoc